Projects/App_003_001/requirements.txt
Python==3.10.12
matplotlib==3.5.1
numpy==1.24.4
pandas==2.0.3
pillow==9.0.1
seaborn==0.11.2
streamlit==1.26.0